Rename classes using PascalCase, change Rentals to Rental
