+++ lib/BaseDecorator.rb
