+++ lib/TrimmerDecorator.rb
@@ -1,4 +1,4 @@
-require_relative 'base_decorator'
+require_relative 'BaseDecorator'
 
 class TrimmerDecorator < Decorator
   def correct_name
