+++ lib/models/Rental.rb
